Add layer of abstraction `FmtWord` between `FmtUnit` and `FmtLine`
Ran into a logical error where groups of characters with different
formatting not separated by a space would get separated by a space in
the output image.

For example, the input "...*Hello*!" would appear as "... Hello !"
(instead of "...Hello!").

The three formatted units "...", "Hello", and "!" are now grouped
together under FmtWord, which defines groups of FmtUnits which should
not be separated by a space.