Finalize `autoRescale()`
Set a maximum default pixel height of 3000, which can be superseded if
the user provides a value. The minimum default is set based on whatever
height is needed to support the default text height of 16, given that
doing so doesn't exceed the maximum (in which case the image height is
set to 3000).