Integrate parsed toml file input with existing caption generation
The parsed and verified toml input is set as a global variable which the
existing code now reads from and writes to when necessary. All
functionality specified by the toml file is also implemented. Namely:
transparency, credits, an option to control image quality, and the option
to output the caption as "parts" (i.e., to split the text and art into
separate images).