Construct FormattedLines by passing in a complete list of FmtWords
Prior in "wrapRegion()", a FormattedLine would be constructed with the
a list of zero or one FmtWords. The FormattedLine's list would then be
modified directly to build up words for that line.

This refactor gives "__init__()" the complete contents of the line as
it's constructed, allowing for a future refactor to change how the
contents are stored internally.